fixed tests and moved jndi binding to JPA only
