+++ persistence-jdbc/core/src/main/resources/play/reference-overrides.conf
@@ -2,6 +2,10 @@ jdbc-journal.slick {
   profile = ${?jdbc-defaults.slick.profile}
   # for backward compatibility only, will be removed in 1.5.0 (@deprecated)
   driver = ${?jdbc-defaults.slick.driver}
+
+  # need to unset jndiName, we must force the akka-persistence-jdbc plugin
+  # to always use the pre-configured DB with AsyncExecutor
+  jndiName = ""
   jndiDbName = ${jdbc-defaults.slick.jndiDbName}
 }
 
@@ -9,6 +13,10 @@ jdbc-read-journal.slick {
   profile = ${?jdbc-defaults.slick.profile}
   # for backward compatibility only, will be removed in 1.5.0 (@deprecated)
   driver = ${?jdbc-defaults.slick.driver}
+
+  # need to unset jndiName, we must force the akka-persistence-jdbc plugin
+  # to always use the pre-configured DB with AsyncExecutor
+  jndiName = ""
   jndiDbName = ${jdbc-defaults.slick.jndiDbName}
 }
 
@@ -16,6 +24,10 @@ jdbc-snapshot-store.slick {
   profile = ${?jdbc-defaults.slick.profile}
   # for backward compatibility only, will be removed in 1.5.0 (@deprecated)
   driver = ${?jdbc-defaults.slick.driver}
+
+  # need to unset jndiName, we must force the akka-persistence-jdbc plugin
+  # to always use the pre-configured DB with AsyncExecutor
+  jndiName = ""
   jndiDbName = ${jdbc-defaults.slick.jndiDbName}
 }
