+++ persistence-jdbc/core/src/main/resources/reference.conf
@@ -25,8 +25,6 @@ jdbc-defaults.slick {
 
 
 db.default {
-  # Configure the default database to be bound to JNDI
-  jndiName=DefaultDS
 
   hikaricp {
     minimumIdle = ${jdbc-defaults.slick.async-executor.minConnections}
