+++ persistence-jdbc/javadsl/src/test/scala/com/lightbend/lagom/javadsl/persistence/jdbc/JdbcPersistenceSpec.scala
@@ -17,6 +17,7 @@ import scala.concurrent.duration._
 import scala.util.Random
 import com.lightbend.lagom.internal.javadsl.persistence.jdbc.SlickProvider
 import com.lightbend.lagom.internal.persistence.ReadSideConfig
+import com.lightbend.lagom.internal.persistence.jdbc.{ SlickDbProvider, SlickDbTestProvider }
 
 abstract class JdbcPersistenceSpec(_system: ActorSystem) extends ActorSystemSpec(_system) {
 
@@ -37,6 +38,7 @@ abstract class JdbcPersistenceSpec(_system: ActorSystem) extends ActorSystemSpec
       val dbName = s"${system.name}_${Random.alphanumeric.take(8).mkString}"
 
       val db = Databases.inMemory(dbName, config = Map("jndiName" -> "DefaultDS"))
+      SlickDbTestProvider.buildAndBindSlickDb(db.dataSource)
       _database = Some(db)
       db
   }
