+++ persistence-jpa/javadsl/src/main/resources/reference.conf
@@ -18,4 +18,10 @@ lagom.persistence.jpa {
 }
 //#persistence
 
+
+db.default {
+  # Configure the default database to be bound to JNDI
+  jndiName = DefaultDS
+}
+
 play.modules.enabled += com.lightbend.lagom.javadsl.persistence.jpa.JpaPersistenceModule
